+++ src/application/contracts/i_wallet_service.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from decimal import Decimal
 from src.infrastructure.database.models.wallet import Wallet
 
 
@@ -9,11 +10,11 @@ class IWalletService(ABC):
         raise NotImplementedError
 
     @abstractmethod
-    async def deposit(self, wallet_id: str, amount: float) -> Wallet:
+    async def deposit(self, wallet_id: str, amount: Decimal) -> Wallet:
         raise NotImplementedError
 
     @abstractmethod
-    async def withdraw(self, wallet_id: str, amount: float) -> Wallet:
+    async def withdraw(self, wallet_id: str, amount: Decimal) -> Wallet:
         raise NotImplementedError
 
     @abstractmethod
